Add automated video downloading Videos are now downloaded and stored in the path given by the "download_base_path" variable in `main.py`. Videos will be saved to "{download_base_path}/{module_code_and_semester}/{video_name}/video.mp4".
As video downloading is currently inefficient due to downloading
multiple streams merged in a single video, future commits are planned
for the user to select only certain streams to download.

For modules already listed in Panopto, simply expand the folders
containing the modules (e.g. the LumiNUS folder) when prompted, then
press Enter in the console.

For other modules, manually log in to LumiNUS with your credentials,
then select a link to a video in the Lecture Videos tab. Afterwards,
click on the module title to be brought to the video selection page
where the module list can be accessed. Click on "Browse" and expand the
folders containing the desired module.